Added time correction to Android client, while doing formatDate(). Although I was able to get a somewhat accurate (time from epoch) for c++, it was still consistently off by around 3571679 milliseconds from Java's System.currentTimeMillis(), which is about 59.52 min. Which is BAD.
Thus, I averaged a bunch of run's times together. This value, called CPP_JAVA_TIME_CORRECTION,
will be added on to diff, so that we can get as close as possible to the actual difference in time.
Since I was testing and sent message immediately, ideally, diff should be very small.

However, formatDate() will be used on older messages too.