Fixed issue of date formatting
The problem is c++'s janky date formatting. It generates it like this:
   Thu Feb  2 23:56:31 2017            //day < 10

...which is janky enough, because of the extra space in between "Feb" and "2", which completely
throws off my delimiter. But also, because this just happened.
   Mon Feb 13 14:33:02 2017            //day > 10

So it turns out, if the day > 10, THERE ISN'T another space. So I had to reaccount for that,
as this was crashing my build.

Also had to do this in ClientOutgoingAsyncTask.

And for some reason, messages sent in ClientOutgoingAsyncTask are getting re-received by my Handler????