Fixed the issue where sending a message also makes you receive it. It turns out, it was an issue with the Socket connections. You see, I didn't really think it through when designing it, but since Socket connections are long term, you should probably send/receive on ONE SOCKET connection, rather than create seperate ones each time you send/receive. Silly me.
By creating 2 socket connections, it was causing my "send" socket to be perceived as another user
on the server; thus, it was getting re-broadcast to me. It had nothing to do with the Message class.

Thus, what I did was create a global static Utilities class to house the Socket object,
called EstablishedSocketConnection.java.

IncomingMessageRunnable and ClientOutgoingAsyncTask now first attempt to get the Socket object
from this class FIRST, before creating their own. They only create it, if the object from the Socket
object class is null